refactor: ConverImageFileSize 클래스로 fileSize 책임 분리
